requirements.txt
azure-functions==1.19.0
Levenshtein==0.25.1
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
python-Levenshtein==0.25.1
rapidfuzz==3.9.1
